requirements.txt
et-xmlfile==1.1.0
openpyxl==3.1.2
packaging==24.0
pdf2image==1.17.0
pillow==10.3.0
PyPDF2==3.0.1
pytesseract==0.3.10
